fix(grouped-polygons): Doesn't break rings
When we were batch rendering polygons and reached MAX_VERTICES_PER_DRAW_CALL vertices rendered, we started a new path. If that happened between the outer and inner rings of a polygon, holes were discarded, resulting in holes being filled.
This avoids breaking polygons in two, solving that issue while keeping the benefits of batching polygons